Add a feature for handlebars rendering
- feature: Allow building without handlebars if desired
