+++ Cargo.toml
@@ -11,7 +11,7 @@ keywords    = ["git", "changelog", "report", "project", "status"]
 categories  = ["command-line-utilities", "development-tools"]
 
 [features]
-default = ["bin"]
+default = ["bin", "handlebars"]
 bin = ["console", "clap", "env_logger"]
 
 [dependencies]
@@ -22,7 +22,7 @@ chrono       = "0.4"
 regex        = "1.4"
 serde        = "1.0"
 console      = { version = "0.14", optional = true }
-handlebars   = "3.5"
+handlebars   = { version = "3.5", optional = true }
 env_logger   = { version = "0.8", optional = true }
 serde_yaml   = "0.8"
 serde_json   = "1.0"
